- Remove unused CSS files and update responsive styling in index.css - make background more aesthetic
