--- frontend/src/components/Admin/ManageProducts/ManageProducts.css
@@ -1,8 +1,0 @@
-@media (max-width: 1024px) {
-    td[data-label]:before {
-        content: attr(data-label);
-        font-weight: bold;
-        display: block;
-        margin-top: 8px;
-    }
-}
