+++ frontend/src/components/Customer/OrderHistory.jsx
@@ -1,7 +1,7 @@
 import { useContext, useEffect, useState } from "react";
 import useAxiosSecure from "../../hooks/useAxiosSecure";
 import { AuthContext } from "../../Auth/AuthProvider";
-import OrderTable from "../shared/OrderTable/OrderTable";
+import OrderTable from "../shared/OrderTable";
 import Loader from "../shared/Loader";
 
 const OrderHistory = () => {
@@ -30,7 +30,7 @@ const OrderHistory = () => {
     }
         
     return (
-        <div className="mt-32">
+        <div className="pt-32 pb-10 glass min-h-dvh">
             <h1 className="text-3xl text-center font-bold text-white">Order History</h1>
             <OrderTable orders={orderHistory} userType={userType}/>
         </div>
