+++ frontend/src/components/shared/SearchBar.jsx
@@ -43,13 +43,13 @@ const SearchBar = ({setAllProducts, setLoading}) => {
 
     return (
         <div
-            className="flex justify-center lg:flex-row  items-center gap-4 relative"
+            className="flex justify-center lg:flex-row  items-center gap-4 relative "
             ref={searchRef}
         >
             <input
                 type="text"
                 placeholder="Search Products"
-                className="p-2 rounded-lg border-2 bg-gray-100 text-primary"
+                className="p-2 rounded-lg bg-gray-100 text-primary border-2 border-primary"
                 value={searchParam}
                 onChange={(e) => setSearchParam(e.target.value)}
             />
