+++ frontend/src/hooks/useAxiosSecure.jsx
@@ -5,8 +5,8 @@ import { AuthContext } from "../Auth/AuthProvider";
 
 
 const axiosSecure = axios.create({
-    baseURL: 'http://localhost:8801',
-    // baseURL: 'https://backend-navy-delta.vercel.app',
+    // baseURL: 'http://localhost:8801',
+    baseURL: 'https://backend-navy-delta.vercel.app',
 });
 
 const useAxiosSecure = () => {
